fixed customer id generation issue

+++ docs/test_dump.sql
@@ -2183,7 +2183,7 @@ CREATE TABLE IF NOT EXISTS `om_tariffs` (
 CREATE TABLE IF NOT EXISTS `om_users` (
   `id` int(11) NOT NULL AUTO_INCREMENT,
   `login` varchar(255) NOT NULL,
-  `customerid` int(11) NOT NULL,
+  `customerid` bigint(20) NOT NULL,
   `basetariffid` int(11) DEFAULT NULL,
   `bundletariffs` varchar(255) DEFAULT NULL,
   `active` int(11) DEFAULT NULL,
